refactor: form/room 폴더 생성
