+++ on-fe/app/room/create/page.jsx
@@ -3,7 +3,7 @@
 import React from 'react'
 import DefaultPageFrame from '@/components/DefaultPageFrame'
 import FormLayOut from '@/components/form/FormLayOut'
-import CreateRoomForm from '@/components/form/CreateRoomForm'
+import CreateRoomForm from '@/components/form/room/CreateRoomForm'
 import BaseModal from '@/components/modal/BaseModal'
 
 const page = () => {
